+++ R/scheduler.R
@@ -35,7 +35,7 @@
 #' \dontrun{
 #' # create a Dockerfile of your script
 #' if(!require(containeRit)){
-#'   devtools::install_github("o2r-project/containerit")
+#'   remotes::install_github("o2r-project/containerit")
 #'   library(containeRit)
 #' }
 #'